fix docs

+++ mindspore/python/mindspore/ops/operations/random_ops.py
@@ -807,7 +807,7 @@ class UniformCandidateSampler(PrimitiveWithInfer):
 
     Examples:
         >>> sampler = ops.UniformCandidateSampler(1, 3, False, 4, 1)
-        >>> output1, output2, output3 = sampler(Tensor(np.array([[1], [3], [4], [6], [3]], dtype=np.int32)))
+        >>> output1, output2, output3 = sampler(Tensor(np.array([[1], [3], [4], [6], [3]], dtype=np.int64)))
         >>> print(output1.shape)
         (3,)
         >>> print(output2.shape)
